// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      places: {
        Row: {
          address: string
          category: Database["public"]["Enums"]["place_category"]
          created_at: string | null
          description: string
          facilities: string[] | null
          id: string
          image: string
          is_favorite: boolean | null
          name: string
          opening_hours: string | null
          rating: number | null
          ticket_price: string | null
          updated_at: string | null
          visited: boolean | null
          visited_date: string | null
        }
        Insert: {
          address: string
          category: Database["public"]["Enums"]["place_category"]
          created_at?: string | null
          description: string
          facilities?: string[] | null
          id?: string
          image: string
          is_favorite?: boolean | null
          name: string
          opening_hours?: string | null
          rating?: number | null
          ticket_price?: string | null
          updated_at?: string | null
          visited?: boolean | null
          visited_date?: string | null
        }
        Update: {
          address?: string
          category?: Database["public"]["Enums"]["place_category"]
          created_at?: string | null
          description?: string
          facilities?: string[] | null
          id?: string
          image?: string
          is_favorite?: boolean | null
          name?: string
          opening_hours?: string | null
          rating?: number | null
          ticket_price?: string | null
          updated_at?: string | null
          visited?: boolean | null
          visited_date?: string | null
        }
        Relationships: []
      }
      reviews: {
        Row: {
          author: string
          comment: string
          created_at: string | null
          date: string
          id: string
          place_id: string
          rating: number
        }
        Insert: {
          author: string
          comment: string
          created_at?: string | null
          date: string
          id?: string
          place_id: string
          rating: number
        }
        Update: {
          author?: string
          comment?: string
          created_at?: string | null
          date?: string
          id?: string
          place_id?: string
          rating?: number
        }
        Relationships: [
          {
            foreignKeyName: "reviews_place_id_fkey"
            columns: ["place_id"]
            isOneToOne: false
            referencedRelation: "places"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      place_category: "cafe" | "restaurant" | "mall" | "historical_place"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never